chore: removed old lib and renamed lib.v2 back to lib
