+++ packages/lit-client/src/index.ts
@@ -1,1 +1,1 @@
-export * from './lib.v2';
+export * from './lib';
