--- packages/lit-client/src/lib.v2/index.ts
@@ -1,1 +1,0 @@
-export * from './LitClient/createLitClient';
