+++ packages/lit-client/src/lib/LitClient/helper/handleNodePromises.ts
