+++ packages/lit-client/src/lib/LitClient/orchestrateHandshake.ts
