+++ packages/lit-client/src/lib/LitNodeClient/helper/most-common-value.ts
