--- packages/lit-client/src/lib/api.bak/index.ts
@@ -1,3 +1,0 @@
-// This folder will contain the modules that 'glue together' chain and lit-node-specific behaviours into our top-level `lit-client` interface
-
-export {};
