--- packages/lit-client/src/lib/chain.bak/index.ts
@@ -1,4 +1,0 @@
-// Define behaviours / methods / modules that are entirely specific to interacting with LIT blockchains
-// These should be 'thin' methods that basically orchestrate a provided `LitNetwork` + params, and use an internal `LitChainClient` to do what is required
-
-export {};
