+++ packages/lit-client/src/lib/index.ts
@@ -1,3 +1,1 @@
-import * as api from './api.bak';
-
-export { api };
+export * from './LitClient/createLitClient';
