--- packages/lit-client/src/lib/lit-nodes.bak/index.ts
@@ -1,4 +1,0 @@
-// Define behaviours / methods / modules that are entirely specific to interacting with LIT nodes
-// These should be 'thin' methods that basically orchestrate a provided `LitNetwork` + params, and use an internal `LitNodeClient` to do what is required
-
-export {};
